requirement.txt
Flask
pymongo
plotly
python-dotenv
